Fix improperly loaded modules
Modules where calling `load_module` with the same name of `__init__`.
This name needs to be unique, so when `clang` module was called first,
and `applellvm` second, under the same name it actually caused
`clang` module under name of `__init__` to reload as `applellvm`.
Ideally, this name will need to include parent package, such as
`matchers` if more plugins will come, but for now, it's enough.